check commit description
the final result is incorrect as there's inconsistency in the way empty layers are represented. for some reason the colours keep disappearing and i'm left with just empty layers. i'm too tired to debug/figure out the reason now so i'll check it again as soon as i can. gnight lol